Fixed regression introduced with PR REFACTOR: File utils new directory by @Samuelopez-ansys in #5851
